+++ src/Imagick/Drawer.php
@@ -15,6 +15,7 @@ use Imagine\Draw\DrawerInterface;
 use Imagine\Exception\InvalidArgumentException;
 use Imagine\Exception\RuntimeException;
 use Imagine\Image\AbstractFont;
+use Imagine\Image\Box;
 use Imagine\Image\BoxInterface;
 use Imagine\Image\Palette\Color\ColorInterface;
 use Imagine\Image\Point;
@@ -125,6 +126,16 @@ final class Drawer implements DrawerInterface
         }
 
         return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @see \Imagine\Draw\DrawerInterface::circle()
+     */
+    public function circle(PointInterface $center, $radius, ColorInterface $color, $fill = false, $thickness = 1)
+    {
+        return $this->ellipse($center, new Box($radius, $radius), $color, $fill, $thickness);
     }
 
     /**
@@ -274,6 +285,43 @@ final class Drawer implements DrawerInterface
         }
 
         return $this;
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @see \Imagine\Draw\DrawerInterface::rectangle()
+     */
+    public function rectangle(PointInterface $leftTop, PointInterface $rightBottom, ColorInterface $color, $fill = false, $thickness = 1)
+    {
+        $minX = min($leftTop->getX(), $rightBottom->getX());
+        $maxX = max($leftTop->getX(), $rightBottom->getX());
+        $minY = min($leftTop->getY(), $rightBottom->getY());
+        $maxY = max($leftTop->getY(), $rightBottom->getY());
+
+        try {
+            $pixel = $this->getColor($color);
+            $rectangle = new \ImagickDraw();
+            $rectangle->setStrokeColor($pixel);
+            $rectangle->setStrokeWidth(max(1, (int) $thickness));
+
+            if ($fill) {
+                $rectangle->setFillColor($pixel);
+            } else {
+                $rectangle->setFillColor('transparent');
+            }
+
+            $rectangle->rectangle($minX, $minY, $maxX, $maxY);
+            $this->imagick->drawImage($rectangle);
+
+            $pixel->clear();
+            $pixel->destroy();
+
+            $rectangle->clear();
+            $rectangle->destroy();
+        } catch (\ImagickException $e) {
+            throw new RuntimeException('Draw rectangle operation failed', $e->getCode(), $e);
+        }
     }
 
     /**